requirements.txt
openai==0.27.8
python-dotenv==1.0.0
SpeechRecognition==3.8.1
pyttsx3==2.90
numpy==1.24.4
gTTS==2.2.4